chore: simplify API renaming RegistryCredentialsForHostname function
